+++ tests/pyxelresttest.py
@@ -41,6 +41,7 @@ class PyxelRestTest(unittest.TestCase):
             http_methods_test_service,
             content_type_test_service,
             base_path_ending_with_slash_test_service,
+            async_test_service,
         )
         serviceshandler.start_services(
             (usual_parameters_test_service, 8943),
@@ -56,6 +57,7 @@ class PyxelRestTest(unittest.TestCase):
             (http_methods_test_service, 8955),
             (content_type_test_service, 8956),
             (base_path_ending_with_slash_test_service, 8957),
+            (async_test_service, 8958),
         )
 
     def test_string_array_parameter(self):
@@ -118,7 +120,7 @@ class PyxelRestTest(unittest.TestCase):
         self.assertRegexpMatches(headers[1][session_header_index], '\d\d\d\d-\d\d-\d\dT\d\d:\d\d:\d\d.\d\d\d\d\d\d')
 
         user_agent_index = headers[0].index('User-Agent')
-        self.assertEqual(headers[1][user_agent_index], 'PyxelRest v0.64.1')
+        self.assertEqual(headers[1][user_agent_index], 'PyxelRest v0.65.0')
 
     def test_post_test_form_parameter(self):
         from pyxelrest import pyxelrestgenerator
@@ -384,6 +386,129 @@ class PyxelRestTest(unittest.TestCase):
         self.assertEqual(
             'http://localhost:8957/test/method',
             pyxelrestgenerator.base_path_ending_with_slash_test_delete_test_method()
+        )
+
+    def test_get_async_url(self):
+        from pyxelrest import pyxelrestgenerator
+        self.assertEqual(
+            {'Status URL': 'http://localhost:8958/test/async/status'},
+            pyxelrestgenerator.async_test_get_test_async()
+        )
+
+    def test_get_custom_url(self):
+        from pyxelrest import pyxelrestgenerator
+        self.assertEqual(
+            [
+                ['X-Custom-Header1', 'X-Custom-Header2'],
+                ['custom1', 'custom2']
+            ],
+            pyxelrestgenerator.pyxelrest_get_url(
+                'http://localhost:8958/test/async/status',
+                extra_headers=[
+                    ['X-Custom-Header1', 'custom1'],
+                    ['X-Custom-Header2', 'custom2'],
+                ],
+            )
+        )
+
+    def test_delete_custom_url(self):
+        from pyxelrest import pyxelrestgenerator
+        self.assertEqual(
+            [
+                ['X-Custom-Header1', 'X-Custom-Header2'],
+                ['custom1', 'custom2']
+            ],
+            pyxelrestgenerator.pyxelrest_delete_url(
+                'http://localhost:8958/test/unlisted',
+                extra_headers=[
+                    ['X-Custom-Header1', 'custom1'],
+                    ['X-Custom-Header2', 'custom2'],
+                ],
+            )
+        )
+
+    def test_post_custom_url_dict(self):
+        from pyxelrest import pyxelrestgenerator
+        self.assertEqual(
+            [
+                ['key1', 'key2', 'key3'],
+                ['value1', 1, 'value3'],
+            ],
+            pyxelrestgenerator.pyxelrest_post_url(
+                'http://localhost:8958/test/dict',
+                [
+                    ['key1', 'key2', 'key3'],
+                    ['value1', 1, 'value3'],
+                ],
+                extra_headers=[
+                    ['Content-Type', 'application/json'],
+                ],
+                parse_body_as='dict',
+            )
+        )
+
+    def test_post_custom_url_dict_list(self):
+        from pyxelrest import pyxelrestgenerator
+        self.assertEqual(
+            [
+                ['key1', 'key2', 'key3'],
+                ['value1', 1, 'value3'],
+                ['other1', 2, 'other3'],
+            ],
+            pyxelrestgenerator.pyxelrest_post_url(
+                'http://localhost:8958/test/dict',
+                [
+                    ['key1', 'key2', 'key3'],
+                    ['value1', 1, 'value3'],
+                    ['other1', 2, 'other3'],
+                ],
+                extra_headers=[
+                    ['Content-Type', 'application/json'],
+                ],
+                parse_body_as='dict_list',
+            )
+        )
+
+    def test_put_custom_url_dict_list(self):
+        from pyxelrest import pyxelrestgenerator
+        self.assertEqual(
+            [
+                ['key1', 'key2', 'key3'],
+                ['value1', 1, 'value3'],
+                ['other1', 2, 'other3'],
+            ],
+            pyxelrestgenerator.pyxelrest_put_url(
+                'http://localhost:8958/test/dict',
+                [
+                    ['key1', 'key2', 'key3'],
+                    ['value1', 1, 'value3'],
+                    ['other1', 2, 'other3'],
+                ],
+                extra_headers=[
+                    ['Content-Type', 'application/json'],
+                ],
+                parse_body_as='dict_list',
+            )
+        )
+
+    def test_put_custom_url_dict(self):
+        from pyxelrest import pyxelrestgenerator
+        self.assertEqual(
+            [
+                ['key1', 'key2', 'key3'],
+                ['value1', 1, 'value3'],
+            ],
+            pyxelrestgenerator.pyxelrest_put_url(
+                'http://localhost:8958/test/dict',
+                [
+                    ['key1', 'key2', 'key3'],
+                    ['value1', 1, 'value3'],
+                ],
+                extra_headers=[
+                    ['Content-Type', 'application/json'],
+                ],
+                parse_body_as='dict',
+            )
         )